Added start of nginx hosted html page on port 80 and run container
I added an html file containing the current date, upload it with ansible in the server.

Moreover, I have updated the dockerfile so that it now initialize nginx at the default page.

I also updated the main file so that it now stops&prune existing containers before running a new one so that port 80 is freed to host a new container